Check if a date is invalid before returning it as a formatted answer.
The approach is that if moment js finds it to be a valid date, an iso-string,
we format it and show it to the user. If not any other string acting as a date
will be returned to the user as is. The aim of this is not to break how
it currently works. There was an option to use `new js/Date answer` but this
would break in cases which we want to have a date such as "2014 04 25 to 2014 06 13"
which is a range and we want to show it to the user as is.

This is based on https://github.com/moment/moment/issues/1407

Signed-off-by: Njagi Mwaniki <njagi@urbanslug.com>