Handle time ranges in `view-by`
The way to do this is to remove the colon and read the time as an int
seems to work ok given a 24 hour format.
We then work with it as an int cast it to string and then add the
colon back as the third last element in the string.

Signed-off-by: Njagi Mwaniki <njagi@urbanslug.com>